+++ java/tenmo-server/src/main/java/com/techelevator/tenmo/dao/AccountDAO.java
@@ -7,7 +7,7 @@ import com.techelevator.tenmo.model.Account;
 
 public interface AccountDAO {
 
-    Account getAccountByUserId(Long userId);
+    Account getAccountByUserId(int userId);
 
     void updateBalance(Account account);
 }
